Moving majority_location and location_visits to query_specs and associated
