+++ flowclient/flowclient/__init__.py
@@ -47,6 +47,8 @@ from .query_specs import (
     unique_locations_spec,
     most_frequent_location_spec,
     total_active_periods_spec,
+    location_visits_spec,
+    majority_location_spec,
 )
 from . import aggregates
 from .aggregates import (
@@ -67,11 +69,8 @@ from .aggregates import (
     unmoving_counts,
     consecutive_trips_od_matrix,
     trips_od_matrix,
-    location_visits,
-    majority_location,
-    majority_location_spec,  # This doesn't seem to be consistent with the rest of this; should these be in query_specs?
-    location_visits_spec,
 )
+from query_specs import majority_location_spec, location_visits_spec
 
 __all__ = [
     "aggregates",
@@ -105,6 +104,4 @@ __all__ = [
     "unmoving_counts",
     "consecutive_trips_od_matrix",
     "trips_od_matrix",
-    "location_visits",
-    "majority_location",
 ]
