+++ flowclient/flowclient/aggregates.py
@@ -1550,24 +1550,3 @@ def histogram_aggregate(*, connection: Connection, **kwargs) -> APIQuery:
         Histogram aggregate query
     """
     return connection.make_api_query(parameters=histogram_aggregate_spec(**kwargs))
-
-
-def majority_location_spec(*, subscriber_location_weights: Dict):
-    return {
-        "query_kind": "majority_location",
-        "subscriber_location_weights": subscriber_location_weights,
-    }
-
-
-@merge_args(majority_location_spec)
-def majority_location(*, connection: Connection, **kwargs):
-    return connection.make_api_query(parameters=majority_location_spec(**kwargs))
-
-
-def location_visits_spec(*, day_trajectories: List):
-    return {"query_kind": "location_visits", "day_trajectories": day_trajectories}
-
-
-@merge_args(location_visits_spec)
-def location_visits(*, connection: Connection, **kwargs):
-    return connection.make_api_query(parameters=location_visits_spec(**kwargs))
